requirements.txt
blinker==1.9.0
certifi~=2024.12.14
charset-normalizer~=3.4.0
click==8.1.8
Flask==3.1.0
Flask-SQLAlchemy
greenlet~=3.1.0
idna~=3.8
itsdangerous==2.2.0
Jinja2~=3.1.3
MarkupSafe~=2.1.5
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy
typing_extensions==4.12.2
urllib3~=2.2.3
Werkzeug==3.1.3

pip~=25.0
attrs~=24.2.0
distro~=1.9.0
wheel~=0.45.1
PySocks~=1.7.1
protobuf~=4.21.12
pyOpenSSL~=24.3.0
cryptography~=43.0.0
pillow~=10.4.0
docutils~=0.21.2
Sphinx~=8.1.3
filelock~=3.16.1
redis~=4.5.4
ipython~=8.30.0
h2~=4.1.0
psutil~=5.9.8
chardet~=5.2.0
simplejson~=3.19.3
importlib_metadata~=8.5.0
pytest~=8.3.4
psycopg2~=2.9.10
Routes~=2.5.1